+++ benchmark/csharp/Tpch/Program.cs
@@ -16,7 +16,7 @@ namespace Tpch
             {
                 Console.WriteLine("Usage:");
                 Console.WriteLine("\t<spark-submit> --master local");
-                Console.WriteLine("\t\t--class org.apache.spark.deploy.DotnetRunner <path-to-microsoft-spark-jar>");
+                Console.WriteLine("\t\t--class org.apache.spark.deploy.dotnet.DotnetRunner <path-to-microsoft-spark-jar>");
                 Console.WriteLine("\t\tTpch.exe <tpch_data_root_path> <query_number> <num_iterations> <true for SQL | false for functional>");
 
                 return;
