+++ docs/building/ubuntu-instructions.md
@@ -172,7 +172,7 @@ Once you build the samples, you can use `spark-submit` to submit your .NET Core 
      ```bash
      spark-submit \
        [--jars <any-jars-your-app-is-dependent-on>] \
-       --class org.apache.spark.deploy.DotnetRunner \
+       --class org.apache.spark.deploy.dotnet.DotnetRunner \
        --master local \
        <path-to-microsoft-spark-jar> \
        <path-to-your-app-binary> <argument(s)-to-your-app>
@@ -182,7 +182,7 @@ Once you build the samples, you can use `spark-submit` to submit your .NET Core 
      - **[Microsoft.Spark.Examples.Sql.Basic](../../examples/Microsoft.Spark.CSharp.Examples/Sql/Basic.cs)**
          ```bash
          spark-submit \
-         --class org.apache.spark.deploy.DotnetRunner \
+         --class org.apache.spark.deploy.dotnet.DotnetRunner \
          --master local \
          ~/dotnet.spark/src/scala/microsoft-spark-2.3.x/target/microsoft-spark-2.3.x-1.0.0-alpha.jar \
          Microsoft.Spark.CSharp.Examples Sql.Basic $SPARK_HOME/examples/src/main/resources/people.json
@@ -190,7 +190,7 @@ Once you build the samples, you can use `spark-submit` to submit your .NET Core 
      - **[Microsoft.Spark.Examples.Sql.Streaming.StructuredNetworkWordCount](../../examples/Microsoft.Spark.CSharp.Examples/Sql/Streaming/StructuredNetworkWordCount.cs)**
          ```bash
          spark-submit \
-         --class org.apache.spark.deploy.DotnetRunner \
+         --class org.apache.spark.deploy.dotnet.DotnetRunner \
          --master local \
          ~/dotnet.spark/src/scala/microsoft-spark-2.3.x/target/microsoft-spark-2.3.x-1.0.0-alpha.jar \
          Microsoft.Spark.CSharp.Examples Sql.Streaming.StructuredNetworkWordCount localhost 9999
@@ -199,7 +199,7 @@ Once you build the samples, you can use `spark-submit` to submit your .NET Core 
          ```bash
          spark-submit \
          --packages org.apache.spark:spark-sql-kafka-0-10_2.11:2.3.2 \
-         --class org.apache.spark.deploy.DotnetRunner \
+         --class org.apache.spark.deploy.dotnet.DotnetRunner \
          --master local \
          ~/dotnet.spark/src/scala/microsoft-spark-2.3.x/target/microsoft-spark-2.3.x-1.0.0-alpha.jar \
          Microsoft.Spark.CSharp.Examples Sql.Streaming.StructuredKafkaWordCount localhost:9092 subscribe test
@@ -208,7 +208,7 @@ Once you build the samples, you can use `spark-submit` to submit your .NET Core 
          ```bash
          spark-submit \
          --jars path/to/net.jpountz.lz4/lz4-1.3.0.jar,path/to/org.apache.kafka/kafka-clients-0.10.0.1.jar,path/to/org.apache.spark/spark-sql-kafka-0-10_2.11-2.3.2.jar,`path/to/org.slf4j/slf4j-api-1.7.6.jar,path/to/org.spark-project.spark/unused-1.0.0.jar,path/to/org.xerial.snappy/snappy-java-1.1.2.6.jar \
-         --class org.apache.spark.deploy.DotnetRunner \
+         --class org.apache.spark.deploy.dotnet.DotnetRunner \
          --master local \
          ~/dotnet.spark/src/scala/microsoft-spark-2.3.x/target/microsoft-spark-2.3.x-1.0.0-alpha.jar \
          Microsoft.Spark.CSharp.Examples Sql.Streaming.StructuredKafkaWordCount localhost:9092 subscribe test
