+++ docs/building/windows-instructions.md
@@ -204,7 +204,7 @@ Once you build the samples, running them will be through `spark-submit` regardle
      ```powershell
      spark-submit.cmd `
        [--jars <any-jars-your-app-is-dependent-on>] `
-       --class org.apache.spark.deploy.DotnetRunner `
+       --class org.apache.spark.deploy.dotnet.DotnetRunner `
        --master local `
        <path-to-microsoft-spark-jar> `
        <path-to-your-app-exe> <argument(s)-to-your-app>
@@ -214,7 +214,7 @@ Once you build the samples, running them will be through `spark-submit` regardle
      - **[Microsoft.Spark.Examples.Sql.Basic](../../examples/Microsoft.Spark.CSharp.Examples/Sql/Basic.cs)**
          ```powershell
          spark-submit.cmd `
-         --class org.apache.spark.deploy.DotnetRunner `
+         --class org.apache.spark.deploy.dotnet.DotnetRunner `
          --master local `
          C:\github\dotnet-spark\src\scala\microsoft-spark-2.3.x\target\microsoft-spark-2.3.x-1.0.0-alpha.jar `
          Microsoft.Spark.CSharp.Examples.exe Sql.Basic %SPARK_HOME%\examples\src\main\resources\people.json
@@ -222,7 +222,7 @@ Once you build the samples, running them will be through `spark-submit` regardle
      - **[Microsoft.Spark.Examples.Sql.Streaming.StructuredNetworkWordCount](../../examples/Microsoft.Spark.CSharp.Examples/Sql/Streaming/StructuredNetworkWordCount.cs)**
          ```powershell
          spark-submit.cmd `
-         --class org.apache.spark.deploy.DotnetRunner `
+         --class org.apache.spark.deploy.dotnet.DotnetRunner `
          --master local `
          C:\github\dotnet-spark\src\scala\microsoft-spark-2.3.x\target\microsoft-spark-2.3.x-1.0.0-alpha.jar `
          Microsoft.Spark.CSharp.Examples.exe Sql.Streaming.StructuredNetworkWordCount localhost 9999
@@ -231,7 +231,7 @@ Once you build the samples, running them will be through `spark-submit` regardle
          ```powershell
          spark-submit.cmd `
          --packages org.apache.spark:spark-sql-kafka-0-10_2.11:2.3.2 `
-         --class org.apache.spark.deploy.DotnetRunner `
+         --class org.apache.spark.deploy.dotnet.DotnetRunner `
          --master local `
          C:\github\dotnet-spark\src\scala\microsoft-spark-2.3.x\target\microsoft-spark-2.3.x-1.0.0-alpha.jar `
          Microsoft.Spark.CSharp.Examples.exe Sql.Streaming.StructuredKafkaWordCount localhost:9092 subscribe test
@@ -240,7 +240,7 @@ Once you build the samples, running them will be through `spark-submit` regardle
          ```powershell
          spark-submit.cmd 
          --jars path\to\net.jpountz.lz4\lz4-1.3.0.jar,path\to\org.apache.kafka\kafka-clients-0.10.0.1.jar,path\to\org.apache.spark\spark-sql-kafka-0-10_2.11-2.3.2.jar,`path\to\org.slf4j\slf4j-api-1.7.6.jar,path\to\org.spark-project.spark\unused-1.0.0.jar,path\to\org.xerial.snappy\snappy-java-1.1.2.6.jar `
-         --class org.apache.spark.deploy.DotnetRunner `
+         --class org.apache.spark.deploy.dotnet.DotnetRunner `
          --master local `
          C:\github\dotnet-spark\src\scala\microsoft-spark-2.3.x\target\microsoft-spark-2.3.x-1.0.0-alpha.jar `
          Microsoft.Spark.CSharp.Examples.exe Sql.Streaming.StructuredKafkaWordCount localhost:9092 subscribe test
