+++ docs/getting-started/ubuntu-instructions.md
@@ -59,7 +59,7 @@ For detailed instructions, you can see [Building .NET for Apache Spark from Sour
 - Run your app.
     ```shell
     spark-submit \
-    --class org.apache.spark.deploy.DotnetRunner \
+    --class org.apache.spark.deploy.dotnet.DotnetRunner \
     --master local \
     microsoft-spark-2.4.x-<version>.jar \
     dotnet HelloSpark.dll
