+++ src/csharp/Microsoft.Spark.E2ETest/SparkFixture.cs
@@ -113,7 +113,7 @@ namespace Microsoft.Spark.E2ETest
             }
 
             // Build the arguments for the spark-submit.
-            string classArg = "--class org.apache.spark.deploy.DotnetRunner";
+            string classArg = "--class org.apache.spark.deploy.dotnet.DotnetRunner";
             string curDir = AppDomain.CurrentDomain.BaseDirectory;
             string jarPrefix = GetJarPrefix(sparkHome);
             string scalaDir = $"{curDir}{sep}..{sep}..{sep}..{sep}..{sep}..{sep}src{sep}scala";
