+++ src/scala/microsoft-spark-2.3.x/src/main/scala/org/apache/spark/sql/api/dotnet/SQLUtils.scala
@@ -10,7 +10,7 @@ import java.util.{List => JList, Map => JMap}
 
 import org.apache.spark.api.python.{PythonAccumulatorV2, PythonBroadcast, PythonFunction}
 import org.apache.spark.broadcast.Broadcast
-import org.apache.spark.deploy.DotnetRunner
+import org.apache.spark.deploy.dotnet.DotnetRunner
 
 object SQLUtils {
